refactor: move menus.ts and widget.ts to workspace/
Part of #1235.
