+++ packages/base/src/commands/index.ts
@@ -30,7 +30,7 @@ import { addProcessingCommands } from '../features/processing/processingCommands
 import keybindings from '../keybindings.json';
 import { getGeoJSONDataFromLayerSource, downloadFile } from '../tools';
 import { JupyterGISTracker, SYMBOLOGY_VALID_LAYER_TYPES } from '../types';
-import { JupyterGISDocumentWidget } from '../widget';
+import { JupyterGISDocumentWidget } from '../workspace/widget';
 
 const POINT_SELECTION_TOOL_CLASS = 'jGIS-point-selection-tool';
