+++ packages/base/src/index.ts
@@ -4,7 +4,7 @@ export * from './constants';
 export * from './features/layers/layerCreationFormDialog';
 export * from './shared/icons';
 export * from './mainview';
-export * from './menus';
+export * from './workspace/menus';
 export * from './workspace/panels';
 export * from './shared/store';
 export * from './features/stac-browser';
@@ -12,4 +12,4 @@ export * from './features/objectproperties';
 export * from './workspace/toolbar';
 export * from './tools';
 export * from './types';
-export * from './widget';
+export * from './workspace/widget';
