+++ packages/base/src/workspace/menus.ts
@@ -1,8 +1,8 @@
 import { CommandRegistry } from '@lumino/commands';
 import { Menu } from '@lumino/widgets';
 
-import { CommandIDs } from './constants';
-import { rasterIcon } from './shared/icons';
+import { CommandIDs } from '../constants';
+import { rasterIcon } from '../shared/icons';
 
 export const vectorSubMenu = (commands: CommandRegistry) => {
   const subMenu = new Menu({ commands });
