+++ packages/base/src/workspace/widget.ts
@@ -17,9 +17,9 @@ import { MessageLoop } from '@lumino/messaging';
 import { ISignal, Signal } from '@lumino/signaling';
 import { SplitPanel, Widget } from '@lumino/widgets';
 
-import { ConsoleView } from './features/console';
-import { JupyterGISMainViewPanel } from './mainview';
-import { MainViewModel } from './mainview/mainviewmodel';
+import { ConsoleView } from '../features/console';
+import { JupyterGISMainViewPanel } from '../mainview';
+import { MainViewModel } from '../mainview/mainviewmodel';
 
 const CELL_OUTPUT_WIDGET_CLASS = 'jgis-cell-output-widget';
